\name{class_configuration}
\alias{class_configuration}
\docType{data}
\title{ Remora Configuration }
\description{
  Holds the \code{\link[frbf:RemoraConfiguration-class]{RemoraConfiguration}} definition.
}
%\usage{data(class_configuration)}
%\format{
%  The format is:
% chr "RemoraConfiguration"
%}
\details{
 This is part of the frbf implementation and is for internal use only.
}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
\author{ Fernando Martins }
%\examples{
%data(class_configuration)
%## maybe str(class_configuration) ; plot(class_configuration) ...
%}
\keyword{ internal }
