\name{class_distance}
\alias{class_distance}
\docType{data}
\title{ Remora Distance }
\description{
  Holds the \code{\link[frbf:RemoraDistance-class]{RemoraDistance}} definition.
}
%\usage{data(class_distance)}
%\format{
%  The format is:
% chr "RemoraDistance"
%}
\details{
 This is part of the frbf implementation and is for internal use only.
}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
%\examples{
%data(class_distance)
%## maybe str(class_distance) ; plot(class_distance) ...
%}
\keyword{ internal }
