\name{class_kernels}
\alias{class_kernels}
\docType{data}
\title{ Remora Kernels }
\description{
  Holds the \code{\link[frbf:RemoraKernels-class]{RemoraKernels}} definition.
}
%\usage{data(class_kernels)}
%\format{
%  The format is:
% chr "RemoraKernels"
%}
\details{
 This is part of the frbf implementation and is for internal use only.
}
%\source{
%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
%\references{
%  ~~ possibly secondary sources and usages ~~
%}
%\examples{
%data(class_kernels)
%## maybe str(class_kernels) ; plot(class_kernels) ...
%}
\keyword{ internal }
