\name{print.RemoraConfiguration}
\alias{print.RemoraConfiguration}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print remora configuration }
\description{
  Print function for remora configuration. It prints the \code{\link[frbf:RemoraConfiguration-class]{configuration}} in a pretty way.
}
\usage{
print.RemoraConfiguration(x, ...)
}
%- maybe also 'usage' for other objects documented here.
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ \code{\link[frbf:RemoraConfiguration-class]{configuration}} to print }
}
\details{
  Overloads the print function for the \code{\link[frbf:RemoraConfiguration-class]{RemoraConfiguration}} object.
}
\value{
  Does not return any value.
}
%\references{ ~put references to the literature/web site here ~ }
%\author{ ~~who you are~~ }
\note{ This function is part of the frbf implementation and is for internal use only. }
\seealso{
\code{\link[frbf:RemoraConfiguration-class]{RemoraConfiguration}}
print
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }