\name{print.RemoraKernels}
\alias{print.RemoraKernels}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print remora kernels }
\description{
  Print function for remora kernels. It prints the \code{\link[frbf:RemoraKernels-class]{kernels}} in a pretty way.
}
\usage{
print.RemoraKernels(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ \code{\link[frbf:RemoraKernels-class]{kernels}} to print }
}
\details{
  Overloads the print function for the \code{\link[frbf:RemoraKernels-class]{RemoraKernels}} object.
}
\value{
  Does not return any value.
}
%\references{ ~put references to the literature/web site here ~ }
\author{ Fernando Martins }
\note{ This function is part of the frbf implementation and is for internal use only. }
\seealso{
\code{\link[frbf:RemoraKernels-class]{RemoraKernels}}
print
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }