\name{print.RemoraModel}
\alias{print.RemoraModel}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Print remora model }
\description{
  Print function for remora model. It prints the \code{\link[frbf:RemoraModel-class]{model}} in a pretty way.
}
\usage{
print.RemoraModel(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ \code{\link[frbf:RemoraModel-class]{model}} to print }
}
\details{
  Overloads the print function for the \code{\link[frbf:RemoraModel-class]{RemoraModel}} object.
}
\value{
  Does not return any value.
}
%\references{ ~put references to the literature/web site here ~ }
%\author{ ~~who you are~~ }
\note{ This function is part of the frbf implementation and is for internal use only. }
\seealso{
\code{\link[frbf:RemoraModel-class]{RemoraModel}}
print
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ internal }